# paasta-monitoring-api/src/kr/paasta/monitoring-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Monitoring Dashboard API",
        "contact": {},
        "version": "5.8.0"
    },
    "paths": {
...
        "/api/v1/ap/bosh/chart/{uuid}": {
            "get": {
                "description": "BOSH 차트 정보를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AP"
                ],
                "summary": "BOSH 차트 정보 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "36dd3d08-5198-42b6-4130-d0c04479236f",
                        "description": "BOSH의 UUID를 주입한다.",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "15m",
                        "description": "기본 시간 범위를 주입한다.",
                        "name": "defaultTimeRange",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "시작 시간을 주입한다.",
                        "name": "timeRangeFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "종료 시간을 주입한다.",
                        "name": "timeRangeTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1m",
                        "description": "시간 단위를 주입한다.",
                        "name": "groupBy",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseInfo": {
                                            "$ref": "#/definitions/v1.BoshChart"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/ap/bosh/process": {
            "get": {
                "description": "BOSH 프로세스 목록을 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AP"
                ],
                "summary": "BOSH 프로세스 목록 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "36dd3d08-5198-42b6-4130-d0c04479236f",
                        "description": "BOSH의 UUID를 주입한다.",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseInfo": {
                                            "$ref": "#/definitions/v1.BoshProcess"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/cp/cluster/average/{type}": {
            "get": {
                "description": "쿠버네티스 클러스터 리소스 사용 평균을 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CP"
                ],
                "summary": "쿠버네티스 클러스터 리소스 사용 평균 가져오기",
                "parameters": [
                    {
                        "enum": [
                            "pod",
                            "cpu",
                            "memory",
                            "disk"
                        ],
                        "type": "string",
                        "description": "Type 정보를 주입한다.",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
        "/api/v1/cp/cluster/worknode": {
            "get": {
                "description": "쿠버네티스 클러스터 노드 정보를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CP"
                ],
                "summary": "쿠버네티스 클러스터 노드 정보 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ip-10-0-0-20",
                        "description": "파라미터에 대한 Pod 관련 메트릭 정보를 반환한다.",
                        "name": "nodename",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "10.0.0.20:9100",
                        "description": "파라미터에 대햔 CPU, Memory, Disk 관련 메트릭 정보를 반환한다.",
                        "name": "instance",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
...
        "/api/v1/cp/pod/container/list": {
            "get": {
                "description": "쿠버네티스 파드 컨테이너 리스트를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CP"
                ],
                "summary": "쿠버네티스 파드 컨테이너 리스트 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "prometheus-kube-prometheus-stack-1648-prometheus-0",
                        "description": "파드명을 주입한다.",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiHelpers.BasicResponseForm": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "responseInfo": {},
                "responseTime": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
        "v1.BoshChart": {
            "type": "object",
            "properties": {
                "defaulttimerange": {
                    "type": "string"
                },
                "groupby": {
                    "type": "string"
                },
                "isconvertkb": {
                    "type": "boolean"
                },
                "metricData": {
                    "type": "object",
                    "additionalProperties": true
                },
                "metricname": {
                    "type": "string"
                },
                "sqlquery": {
                    "type": "string"
                },
                "timerangefrom": {
                    "type": "string"
                },
                "timerangeto": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "v1.BoshProcess": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "string"
                },
                "memory": {
                    "type": "number"
                },
                "pid": {
                    "type": "string"
                },
                "process": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
...
}